# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HabitFlow API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type HabitFrequency =
  (typeof HabitFrequency)[keyof typeof HabitFrequency];

export const HabitFrequency = {
  daily: "daily",
  weekdays: "weekdays",
  weekends: "weekends",
  custom: "custom",
} as const;

export interface Habit {
  id: number;
  name: string;
  description?: string | null;
  icon: string;
  color: string;
  frequency: HabitFrequency;
  customDays?: number[] | null;
  targetCount: number;
  streak: number;
  longestStreak: number;
  totalCompletions: number;
  createdAt: string;
}

export type CreateHabitRequestFrequency =
  (typeof CreateHabitRequestFrequency)[keyof typeof CreateHabitRequestFrequency];

export const CreateHabitRequestFrequency = {
  daily: "daily",
  weekdays: "weekdays",
  weekends: "weekends",
  custom: "custom",
} as const;

export interface CreateHabitRequest {
  name: string;
  description?: string | null;
  icon: string;
  color: string;
  frequency: CreateHabitRequestFrequency;
  customDays?: number[] | null;
  targetCount: number;
}

export type UpdateHabitRequestFrequency =
  (typeof UpdateHabitRequestFrequency)[keyof typeof UpdateHabitRequestFrequency];

export const UpdateHabitRequestFrequency = {
  daily: "daily",
  weekdays: "weekdays",
  weekends: "weekends",
  custom: "custom",
} as const;

export interface UpdateHabitRequest {
  name?: string;
  description?: string | null;
  icon?: string;
  color?: string;
  frequency?: UpdateHabitRequestFrequency;
  customDays?: number[] | null;
  targetCount?: number;
}

export interface HabitCompletion {
  id: number;
  habitId: number;
  completedDate: string;
  createdAt: string;
}

export interface CompleteHabitRequest {
  date: string;
}

export type HabitStatWeeklyDataItem = {
  date: string;
  completed: boolean;
};

export interface HabitStat {
  habitId: number;
  name: string;
  icon: string;
  color: string;
  streak: number;
  longestStreak: number;
  totalCompletions: number;
  completionRate: number;
  weeklyData: HabitStatWeeklyDataItem[];
}

export interface StatsResponse {
  totalHabits: number;
  completedToday: number;
  longestStreakEver: number;
  insights: string[];
  habitStats: HabitStat[];
}

export type ListCompletionsParams = {
  from: string;
  to: string;
};