# lib/api-zod/src/generated/types/createHabitRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HabitFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateHabitRequestFrequency } from "./createHabitRequestFrequency";

export interface CreateHabitRequest {
  name: string;
  description?: string | null;
  icon: string;
  color: string;
  frequency: CreateHabitRequestFrequency;
  customDays?: number[] | null;
  targetCount: number;
}