# lib/api-zod/src/generated/types/habit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HabitFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { HabitFrequency } from "./habitFrequency";

export interface Habit {
  id: number;
  name: string;
  description?: string | null;
  icon: string;
  color: string;
  frequency: HabitFrequency;
  customDays?: number[] | null;
  targetCount: number;
  streak: number;
  longestStreak: number;
  totalCompletions: number;
  createdAt: string;
}