# lib/api-zod/src/generated/types/habitCompletion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HabitFlow API
 * OpenAPI spec version: 0.1.0
 */

export interface HabitCompletion {
  id: number;
  habitId: number;
  completedDate: string;
  createdAt: string;
}