# lib/api-zod/src/generated/types/habitStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HabitFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { HabitStatWeeklyDataItem } from "./habitStatWeeklyDataItem";

export interface HabitStat {
  habitId: number;
  name: string;
  icon: string;
  color: string;
  streak: number;
  longestStreak: number;
  totalCompletions: number;
  completionRate: number;
  weeklyData: HabitStatWeeklyDataItem[];
}