# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HabitFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { HabitStat } from "./habitStat";

export interface StatsResponse {
  totalHabits: number;
  completedToday: number;
  longestStreakEver: number;
  insights: string[];
  habitStats: HabitStat[];
}