# lib/api-zod/src/generated/types/updateHabitRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HabitFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateHabitRequestFrequency } from "./updateHabitRequestFrequency";

export interface UpdateHabitRequest {
  name?: string;
  description?: string | null;
  icon?: string;
  color?: string;
  frequency?: UpdateHabitRequestFrequency;
  customDays?: number[] | null;
  targetCount?: number;
}